requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==23.2.1
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
